fix all clippy warnings: doc indentation, casts, field init, assertions
Resolve ~90 clippy lints across 33 files so -D warnings passes clean.
doc_lazy_continuation (indent list continuations), field_reassign_with_default
(use struct init syntax), unnecessary_cast, needless_borrow, collapsible_if,
manual_range_contains, bool_assert_comparison, err_expect, single_match,
approx_constant (use std::f64::consts::PI), needless_range_loop, identity_op,
manual_saturating_arithmetic. Allow attributes for too_many_arguments,
type_complexity, and assertions_on_constants where structural.

+++ src/monitor/bpf_map/tests/htab_tests.rs
@@ -1254,7 +1254,7 @@ fn iter_percpu_htab_entries_zero_pptr_returns_empty_per_cpu() {
     map.map_type = BPF_MAP_TYPE_PERCPU_HASH;
     // Overwrite map_type in the buffer too (htab struct read).
     let htab = test_htab_offsets();
-    let map_type_off = (offsets.map_type) as usize;
+    let map_type_off = offsets.map_type;
     buf[map_type_off..map_type_off + 4].copy_from_slice(&BPF_MAP_TYPE_PERCPU_HASH.to_ne_bytes());
     // val at value position is already zero (built with val=0u64).
     let _ = htab;
